# Vsync via the Windows compositor when appropriate.