--- a/platform/windows/context_gl_windows.cpp
+++ b/platform/windows/context_gl_windows.cpp
@@ -34,6 +34,8 @@
 
 #include "context_gl_windows.h"
 
+#include <dwmapi.h>
+
 #define WGL_CONTEXT_MAJOR_VERSION_ARB 0x2091
 #define WGL_CONTEXT_MINOR_VERSION_ARB 0x2092
 #define WGL_CONTEXT_FLAGS_ARB 0x2094
@@ -63,16 +65,72 @@ int ContextGL_Windows::get_window_height() {
 	return OS::get_singleton()->get_video_mode().height;
 }
 
+bool ContextGL_Windows::should_vsync_via_compositor() {
+
+	if (OS::get_singleton()->is_window_fullscreen()) {
+		return false;
+	}
+
+	// Note: All Windows versions supported by Godot (Vista and later)
+	// have a compositor.  It can be disabled on earlier Windows versions.
+	BOOL dwm_enabled;
+
+	if (SUCCEEDED(DwmIsCompositionEnabled(&dwm_enabled))) {
+		return dwm_enabled;
+	}
+
+	return false;
+}
+
 void ContextGL_Windows::swap_buffers() {
+
+	if (use_vsync) {
+		bool vsync_via_compositor_now = should_vsync_via_compositor();
+
+		if (vsync_via_compositor_now) {
+			DwmFlush();
+		}
+
+		if (vsync_via_compositor_now != vsync_via_compositor) {
+
+			// One of the following just occurred...
+			//
+			// 1) Vertical syncing via the compositor has just been enabled.
+			//    This will happen in the following cases:
+			//      a) When switching from full screen mode to windowed mode
+			//         and the OS compositor is enabled.
+			//      b) When, in windowed mode, the user enables the compositor--
+			//         by switching themes, for example.
+			//
+			//    In this case turn vsync via the compositor on and disable it
+			//    for OpenGL.
+			//
+			// 2) Vertical syncing via the compositor has just been disabled.
+			//    This will happen in the following cases:
+			//      a) When switching from windowed mode to full screen mode
+			//         regardless whether the OS compositor is enabled or not.
+			//      b) When, in windowed mode, the user disables the
+			//         compositor--by switching themes, for example.
+			//
+			//    In this case turn vsync via OpenGL on and disable it for the
+			//    compositor.
+
+			set_use_vsync(true);
+		}
+	}
 
 	SwapBuffers(hDC);
 }
 
 void ContextGL_Windows::set_use_vsync(bool p_use) {
 
+	vsync_via_compositor = p_use && should_vsync_via_compositor();
+
 	if (wglSwapIntervalEXT) {
-		wglSwapIntervalEXT(p_use ? 1 : 0);
+		int swap_interval = (p_use && !vsync_via_compositor) ? 1 : 0;
+		wglSwapIntervalEXT(swap_interval);
 	}
+
 	use_vsync = p_use;
 }
 
@@ -177,6 +235,7 @@ ContextGL_Windows::ContextGL_Windows(HWND hwnd, bool p_opengl_3_context) {
 	opengl_3_context = p_opengl_3_context;
 	hWnd = hwnd;
 	use_vsync = false;
+	vsync_via_compositor = false;
 }
 
 ContextGL_Windows::~ContextGL_Windows() {