@TerminalJack 2019-10-28
This is a resubmitted PR.  The original is [here](https://github.com/godotengine/godot/pull/33127).