@lawnjelly 2019-10-29
Is it possible to provide more information on why this should help with jitter, and by what mechanisms?