@starry-abyss 2019-10-29
My speculative and simplified vision of the compositing is as follows:

ideal:
1) Godot renders in the window
2) Firefox renders in the window
3) Compositor waits for v-blank and renders the mix of all windows
4) We have current frame of Godot shown

practical:
1) Godot waits for v-blank
2) Firefox renders in the window
3) Compositor waits for v-blank and renders the mix of all windows
4) Depending on whether compositor or Godot was faster, we have current or previous frame of Godot shown

DwmFlush is one of the ways to make sure Godot always renders before the compositor. There are sources claiming using it as wrong (e.g. https://www.vsynctester.com/firefoxisbroken.html#asolution), but IMHO it's way better than just relying on luck.