@lawnjelly 2019-10-29
> I wish I understood the problem better. As it is, this is a case of monkey see, monkey do. Some of the posters in the issues threads suggested that someone look at other projects to see what they do about the issue and this is what [glfw](https://github.com/glfw/glfw/blob/51bb76c7c3de934b2dd3affcdc297a32e4817835/src/wgl_context.c) and [Chromium](https://cs.chromium.org/chromium/src/third_party/glfw/src/src/wgl_context.c?q=dwmflush()&sq=package:chromium&g=0&l=265) do.

I was kind of getting that impression. Don't get me wrong I agree if there is a bug in the compositor on windows in some configurations and we need to work around it we should, and this may well be a good solution. But rather than merge this straight away I think it would be good to do a little more research to confirm if this is the best fix (and try and work out why it works, if it does work!) :smile: .

Having the frame present at an unpredictable time isn't the direct problem (its not ideal), the problems really come because the game simulation needs some kind of idea of the presentation time in order to simulate objects in the correct positions. This is a major flaw in the graphics APIs currently, and is even best case a guessing game on desktop (although there is a vulkan extension in the works to help with this I believe). If the estimation of the frame time is off by varying amounts, you get jitter. If the estimation is off, but by a fairly regular gap, you get smoother playback.

See #30791 and also the links within, particularly:
https://medium.com/@alen.ladavac/the-elusive-frame-timing-168f899aec92

The vsynctester article is useful, he mentions some other features of the DWM_TIMING_INFO structure which may be an alternative? qpcVBlank for instance. He also mentions that chrome uses a timer.

I already have a fork of Godot that adds delta smoothing, this may be equivalent to the effect chrome is using. The reason I suggested in the issue to try using `--fixed-fps 60` in the command line, is that we can test the effect of fixing the delta input on such a system. This will help understand whether the jitter is caused by the input timings, or due to random delays in the output frame (much as starry-abyss suggests above).

If the command line gives less jitter, the problem is at least partly due to input timings and would be helped by delta smoothing (which is something all platforms can benefit from). If the command line gives the same amount or more jitter, then that gives further weight to either using the dwmFlush fix. It would be a good idea to find this out, because from the article it appears as though dwmFlush will introduce timing discrepancies itself.