@TerminalJack 2019-10-29
@lawnjelly Yes, I agree that we want to test and understand this as well as we can before merging it since the change is in a fairly critical path for both the (Windows) editor and the end-user's game.

Assuming that those other projects really do use this same strategy then I'd say we are in pretty good company.

From my testing, the call to `DwmFlush()` blocks the thread until the vertical blank period and returns once the compositor is done with its work.  For the game I was testing it will return about 1ms to 2ms after the vertical blank.

This is what `SwapBuffers()` does as well (when it works correctly) when using a swap interval of 1 (double buffering, syncing to vertical blank.)  The only difference between the two that I can see is that `SwapBuffers()` takes longer to run.  It doesn't return until 3ms to 4ms after the vertical blank.  

So, in theory, you get a little extra processing time and less latency by using the compositor for vsync.

Later today I will post some results of the timing regarding `DwmFlush()` and `SwapBuffers()` in relation to the vertical blank period.  I'll do this for when `SwapBuffers()` is behaving as it should (when in full screen) and when it is not (old code, windowed mode.)