@starry-abyss 2019-10-29
> This is a major flaw in the graphics APIs currently

What I'm experiencing is the sprite goes back and forth, by ~1/3 of its size. Either from time to time (stutter?), or over and over again like mad - jitter. This only ever happens with Godot.
It would be cool if microscopic imprecisions are solved as well, but currently the engine produces severely broken visuals.