@lawnjelly 2019-10-29
> From my testing, the call to `DwmFlush()` blocks the thread until the vertical blank period and returns once the compositor is done with its work. For the game I was testing it will return about 1ms to 2ms after the vertical blank.

How are you measuring the time of the vertical blank?