@TerminalJack 2019-10-29
I was using `DwmGetCompositionTimingInfo()`.  It will tell you when the next vertical blank period is.

I was basically using some version of the following code:

	static void log_vblank_info() {
		LARGE_INTEGER now;
		QueryPerformanceCounter(&now);

		DWM_TIMING_INFO timing_info = { sizeof(timing_info) };

		DwmGetCompositionTimingInfo(NULL, &timing_info);

		LARGE_INTEGER frequency;
		QueryPerformanceFrequency(&frequency);

		int64_t now_us     = now.QuadPart * 1000000 / frequency.QuadPart;
		int64_t vblank_us  = timing_info.qpcVBlank * 1000000 / frequency.QuadPart;
		// int64_t compose_us = timing_info.qpcCompose * 1000000 / frequency.QuadPart;

		std::stringstream msg;
		//msg << "now_us = " << now_us << ", vblank_us = " << vblank_us;
		//msg << "now_us = " << now_us << ", compose_us = " << compose_us;
		msg << std::setprecision(5) << double(vblank_us - now_us) / 1000.0;

		log_msg(msg.str());
	}

`log_msg()` uses a background thread to write the message to file.

`log_vblank_info()` was being called in `ContextGL_Windows::swap_buffers()`...

	void ContextGL_Windows::swap_buffers() {

		...

		SwapBuffers(hDC);

		if (use_vsync) {
			log_vblank_info();
		}
	}


After the file was created I had to sort the data out but the gist of it was as I stated earlier.