@lawnjelly 2019-10-29
I think you should try using the qpcVBlank value or the qpcCompose value to pass to main_timer_sync in main::iteration.

If you look in `main/main.cpp` line 1906 you will see this:

```
	uint64_t ticks = OS::get_singleton()->get_ticks_usec();
	Engine::get_singleton()->_frame_ticks = ticks;
	main_timer_sync.set_cpu_ticks_usec(ticks);
	main_timer_sync.set_fixed_fps(fixed_fps);
```
This is the main frame time that gives the delta which drives godot. (Yes you can smooth it close to here, I do it in main_timer_sync!).

Now instead of randomly grabbing whatever the current time is when this is reached, try substituting in the value (converted to usec) from qpcVBlank etc from DWM_TIMING_INFO.

For a test you can just create a global, store this in your log_vblank_info() cpp, then extern this in main.cpp and load it directly instead of the call to OS::get_singleton()->get_ticks_usec().

This could well give far better deltas.